+++ tests/test_search_request.py
@@ -12,3 +12,17 @@ def test_search_request():
         start_index=1,
         count=10,
     )
+
+
+def test_count_floor():
+    """Test that count values less than 1 are interpreted as 1.
+    https://datatracker.ietf.org/doc/html/rfc7644#section-3.4.2.4
+
+        A value less than 1 SHALL be interpreted as 1.
+    """
+
+    sr = SearchRequest(count=0)
+    assert sr.count == 1
+
+    sr = SearchRequest(count=-1)
+    assert sr.count == 1