+++ doc/changelog.rst
@@ -1,6 +1,10 @@
 Changelog
 =========
 
+Added
+^^^^^
+- :attr:`~scim2_models.SearchRequest.count` value is floored to 1
+
 [0.1.6] - 2024-06-06
 --------------------
 Added